+++ cart/domain/cart/method.go
@@ -1,0 +1,8 @@
+package cart
+
+type (
+	PaymentMethod struct {
+		Code    string
+		Gateway string
+	}
+)