+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## v3.13.0 [upcoming]
 **cart**
-* **Breaking:** Added a 'Gateway' field to SplitQualifier, to ensure a diversity of payment possibilities.
+* **Breaking:** Added a 'Gateway' field to the payment method information, provided to create a payment selection. The Gateway in PaymentMethod represents an actual payment gateway.
+* **Breaking:** Added a 'Gateway' field to SplitQualifier, representing an actual payment gateway, to ensure a diversity of payment possibilities.
 
 ## v3.12.0 
 **product**